feat: Add Services struct and User interface to service package
This commit adds the Services struct and User interface to the service package. It also includes the implementation of the NewUser function in the user.go file. These changes allow for the creation of user-related services in the application.

Refactor the NewServices function to return the Services struct instead of being void. This change ensures that the Services struct is properly initialized and can be used throughout the application.

The purpose of this commit is to enhance the functionality of the application by introducing the Services struct and User interface, as well as providing a way to create user-related services.