feat: Add columns for icon and color to roles table
This commit adds two new columns, icon and color, to the roles table in the database. The icon column is of type TEXT and has a default value of an empty string, while the color column is also of type TEXT and has a default value of an empty string. These columns will be used to store additional information about each role.